+++ rust/openvasd/src/controller/entry.rs
@@ -6,14 +6,10 @@
 //!
 //! All known paths must be handled in the entrypoint function.
 
-use std::convert::Infallible;
-
 use std::{fmt::Display, sync::Arc};
 
 use super::{context::Context, quit_on_poison};
-use futures_util::Stream;
 use hyper::{Body, Method, Request, Response};
-use serde::Serialize;
 
 use crate::scan::{Error, ScanDeleter, ScanStarter, ScanStopper};
 /// The supported paths of scannerd
@@ -30,63 +26,6 @@ enum KnownPaths {
     Unknown,
 }
 
-struct JsonArrayStreamer<T> {
-    elements: Vec<T>,
-    first: bool,
-}
-
-impl<T> Unpin for JsonArrayStreamer<T> {}
-
-impl<T> JsonArrayStreamer<T>
-where
-    T: Serialize,
-{
-    fn new(elements: Vec<T>) -> Self {
-        Self {
-            elements,
-            first: true,
-        }
-    }
-}
-
-impl<T> Stream for JsonArrayStreamer<T>
-where
-    T: Serialize,
-{
-    type Item = Result<String, Infallible>;
-
-    fn poll_next(
-        mut self: std::pin::Pin<&mut Self>,
-        _: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Option<Self::Item>> {
-        let out = {
-            if self.first && self.elements.is_empty() {
-                self.first = false;
-                Some("[]".to_string())
-            } else if self.first {
-                self.first = false;
-                Some("[".to_string())
-            } else {
-                match self.elements.pop() {
-                    Some(e) => {
-                        let e = serde_json::to_string(&e).unwrap();
-                        if self.elements.is_empty() {
-                            Some(format!("{}]", e))
-                        } else {
-                            Some(format!("{},", e))
-                        }
-                    }
-                    None => None,
-                }
-            }
-        };
-
-        match out {
-            Some(e) => std::task::Poll::Ready(Some(Ok(e))),
-            None => std::task::Poll::Ready(None),
-        }
-    }
-}
 impl KnownPaths {
     #[tracing::instrument]
     /// Parses a path and returns the corresponding `KnownPaths` variant.
@@ -354,12 +293,8 @@ where
             .await
         }
         (&Method::GET, Vts) => {
-            // need to get rid of clone
             let (_, oids) = ctx.oids.read()?.clone();
-
-            let stream = JsonArrayStreamer::new(oids);
-
-            Ok(ctx.response.ok_stream(stream).await)
+            Ok(ctx.response.ok_stream(oids).await)
         }
         _ => Ok(ctx.response.not_found("path", req.uri().path())),
     }
